a couple things
text based menu is no longer case sensitive. 
users can now add priority skills to their character, and can choose any form of Craft, Knowledge, Perform, or Profession, and have the appropriate ability score modifier. If the given priority skill is not recognized as one of these it is designed as Intelligence based.