User can now multiclass with non-core base classes
You can use the levelup command to input any class name you want, and the program will ask you for the base characteristics of the class before creating it and adding it to the DDCharacter's classes HashMap. Special abilities still must be added manually through the addspecial command in the text based menu.